Made the project modular
Moved the commands that describe the font and packages used in separate
files so that it's easier to change the font used or to add additional
packages.

I also moved the commands that describe the project type, name, group
etc. in separate files so that they can be modified more easily.